// frontend/src/types/Apartment.ts
export interface Apartment {
    aptSeq: string;
    sggCd: string;
    umdCd: string;
    umdNm: string;
    jibun: string;
    roadNmSggCd: string;
    roadNm: string;
    roadNmBonbun: string;
    roadNmBubun: string;
    aptNm: string;
    buildYear: number;
    latitude: string;
    longitude: string;
    name?: string;
    latestDealAmount?: string;
    address?: string;
}